Move more code into SoftwareTransformCommon, cleanup
