+++ Windows/GEDebugger/TabVertices.cpp
@@ -170,7 +170,7 @@ int CtrlVertexList::GetRowCount() {
 	auto state = gpuDebug->GetGState();
 
 	int rowCount_ = gpuDebug->GetCurrentPrimCount();
-	if (!gpuDebug->GetCurrentSimpleVertices(rowCount_, vertices, indices)) {
+	if (!gpuDebug->GetCurrentDrawAsDebugVertices(rowCount_, vertices, indices)) {
 		rowCount_ = 0;
 	}
 	VertexDecoderOptions options{};
